+++ CHANGELOG.md
@@ -25,6 +25,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `edit`, `edit-multiple`, `show`, `clone` support "^n" expression to select a time entry to act
   on, "^0" is the same as "current", "^1" is the same as "last", "^2" chooses the time entry before
   the last one, etc.
+- new `md` (markdown) format to print time entries
 
 ## [v0.24.1] - 2021-09-20